% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixoglmm.R
\docType{data}
\name{data_toy}
\alias{data_toy}
\title{Data set toy example}
\format{
A data frame with 500 rows and 8 variables
}
\usage{
data(data_toy)
}
\description{
A simulated data set containing 8 variables (5 responses variables and 3 covariates).
\itemize{
  \item \code{y1} numeric, gaussian response variable
  \item \code{y2} numeric, gaussian response variable
  \item \code{y3} numeric, gaussian response variable
  \item \code{Po1} numeric, response variable with counts simulated from a Poison distribution
  \item \code{Be1} factor with two categories, binary response variable
  \item \code{X1} continuous covariate
  \item \code{X2} continuous covariate
  \item \code{X3} categorical covariate with three levels.
}
}
