% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ranef.R
\name{extract_ranef}
\alias{extract_ranef}
\title{Extract random effects for mixoglmm.}
\usage{
extract_ranef(object, method = c("conditional means",
  "conditional modes"))
}
\arguments{
\item{object}{object of class \code{'mixoglmm'}}

\item{method}{conditional modes or conditional means of the random effects}
}
\description{
extract random effects .
}
