% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixoglmm.R
\name{mixoglmm}
\alias{mixoglmm}
\title{Fitting A Generalized Linear Model For Mixed Responses Of Type Gaussian, Poisson and
    Binomial Using A Shared Random Effect}
\usage{
mixoglmm(formula, data, families = NULL, cor_struct_gauss = NULL,
  constraints.beta = NULL, constraints.lambda = NULL, Ntrials = NULL,
  na.action, weights, offset = NULL, contrasts = NULL,
  control = mixoglmm.control(), ...)
}
\arguments{
\item{formula}{A formula as in the Formula package?}

\item{data}{data}

\item{families}{named list of families. Each element of the list is a family() as in glm. names must correspond to the responses s
pecified in the formula.only gaussian(), binomial() and poisson() are supported so far.}

\item{cor_struct_gauss}{correlation structure of the gaussian responses}

\item{constraints.beta}{list of constraints}

\item{constraints.lambda}{list of constraints of length one}

\item{Ntrials}{list of number of trials to be used for the binomial. Defaults to NULL, will be one.}

\item{na.action}{eliminate NAs or keep them}

\item{weights}{weights which need to be constant across multiple measurements. Negative weights are not allowed.}

\item{offset}{this can be used to specify an a priori known component to be included in the linear predictor during fitting.}

\item{contrasts}{an optional list. See the \code{contrasts.arg} of \code{\link{model.matrix.default}}.}

\item{control}{list of parameters for controlling the fitting process. See \code{\link{mixoglmm.control}} for details.}

\item{...}{additional arguments.}
}
\value{
an object of class mixoglmm
}
\description{
mixoglmm is used to fit a model with shared random effects for normal and nonnormal random variables.
}
\examples{
sum(1:10)

\dontrun{
sum("a")
}
}
