% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixoglmm.R
\name{print.mixoglmm}
\alias{print.mixoglmm}
\title{Print Method for class mixoglmm.}
\usage{
\method{print}{mixoglmm}(x, call = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{'mixoglmm'}}

\item{call}{displays function call if \code{TRUE}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints regression coefficients, variance of random effects and parameters of the error structure of class mixoglmm.
}
