% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixoglmm.R
\name{summary.mixoglmm}
\alias{summary.mixoglmm}
\title{Summary method for mixoglmm.}
\usage{
\method{summary}{mixoglmm}(object, call = FALSE, ...)
}
\arguments{
\item{object}{object of class \code{'mixoglmm'}}

\item{call}{displays function call if \code{TRUE}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary of thresholds, regression coefficients
and parameters of the error structure of class \code{'mixoglmm'}.
}
