% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixoglmm.R
\name{vcov.mixoglmm}
\alias{vcov.mixoglmm}
\title{vcov of Multivariate Ordinal Regression Models.}
\usage{
\method{vcov}{mixoglmm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{'mixoglmm'}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{vcov} is a generic function which extracts the Godambe information matrix from objects of class \cr
\code{'mixoglmm'}.
}
